1. 硬币数存储类型由int更改为double，显示时向下取整保留一位小数 2. 添加了“渔夫”天赋 3. 优化获得AddPoint值和NextLevelNeedCoin值的方式逻辑 4. 修复了上个提交后，升级天赋时抛出错误的问题。 5. 新增TalentGivingItems枚举以存储装备天赋时要给予的物品。6. 新建Mth类及向下取整保留指定小数位数方法

+++ src/main/java/com/ixbob/thepit/event/OnPlayerOwnCoinModifiedListener.java
@@ -11,14 +11,14 @@ public class OnPlayerOwnCoinModifiedListener implements Listener {
     @EventHandler
     public void onPlayerOwnCoinModified(PlayerOwnCoinModifiedEvent event) {
         Player player = event.getPlayer();
-        int newCoin = event.getNewCoin();
+        double newCoin = event.getNewCoin();
         PlayerDataBlock dataBlock = Main.getPlayerDataBlock(player);
         updateDataBlock(dataBlock, newCoin);
         dataBlock.updatePlayerDBData();
         dataBlock.updateScoreboardOwnCoinAmount();
     }
 
-    private void updateDataBlock(PlayerDataBlock dataBlock, int newCoin) {
+    private void updateDataBlock(PlayerDataBlock dataBlock, double newCoin) {
         dataBlock.setCoinAmount(newCoin);
     }
 }
